refactor(test): split oversized wiring and compat migration tests
