+++ internal/app/runtime_control_wiring_test.go
@@ -2,27 +2,20 @@ package app
 
 import (
 	"context"
-	"errors"
-	"fmt"
 	"io"
 	"net"
 	"os"
 	"path/filepath"
 	"strings"
-	"sync/atomic"
-	"syscall"
 	"testing"
 	"time"
 
-	runtimev1 "github.com/bnema/gordon/api/gordon/runtime/v1"
 	"github.com/bnema/gordon/internal/boundaries/out"
 	"github.com/bnema/gordon/internal/domain"
 	"github.com/spf13/viper"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/metadata"
 	"google.golang.org/grpc/status"
 )
 
@@ -62,968 +55,6 @@ type signalCloser chan struct{}
 func (closer signalCloser) Close() error {
 	close(closer)
 	return nil
-}
-
-func TestCreateRuntimeCommandClientUsesPrivateUnixSocketWithRequiredToken(t *testing.T) {
-	client, err := createRuntimeCommandClient(context.Background(), RuntimeControlConfig{
-		Endpoint: "unix:///var/lib/gordon/migration/fixture/runtime-control.sock",
-		Token:    "component-token",
-	})
-	require.NoError(t, err)
-	require.NotNil(t, client)
-	closer, ok := client.(interface{ Close() error })
-	require.True(t, ok)
-	t.Cleanup(func() { require.NoError(t, closer.Close()) })
-	_, err = createRuntimeCommandClient(context.Background(), RuntimeControlConfig{Endpoint: "unix:///var/lib/gordon/migration/fixture/runtime-control.sock"})
-	require.Error(t, err, "Unix transport requires a component token")
-}
-
-func TestCreateRuntimeCommandClientRejectsUnixEndpointOutsideMigrationState(t *testing.T) {
-	_, err := createRuntimeCommandClient(context.Background(), RuntimeControlConfig{
-		Endpoint: "unix:///tmp/runtime-control.sock",
-		Token:    "component-token",
-	})
-	require.Error(t, err, "only the generated migration Unix socket is allowed")
-}
-
-func TestPostHandoffRuntimeClientDialsOnlyValidatedHostMigrationSocket(t *testing.T) {
-	root, err := os.MkdirTemp("/tmp", "gordon-recovery-")
-	require.NoError(t, err)
-	t.Cleanup(func() { require.NoError(t, os.RemoveAll(root)) })
-	path := filepath.Join(root, "migration", "fixture", bootstrapRuntimeSocketName)
-	require.NoError(t, os.MkdirAll(filepath.Dir(path), 0o700))
-	listener, err := net.Listen("unix", path)
-	require.NoError(t, err)
-	t.Cleanup(func() { _ = listener.Close() })
-
-	const token = "generated-recovery-token"
-	server := grpc.NewServer(grpc.UnaryInterceptor(func(ctx context.Context, req any, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (any, error) {
-		if authorization := metadata.ValueFromIncomingContext(ctx, "authorization"); len(authorization) != 1 || authorization[0] != "Bearer "+token {
-			return nil, status.Error(codes.Unauthenticated, "runtime authentication failed")
-		}
-		return handler(ctx, req)
-	}))
-	runtimev1.RegisterRuntimeServiceServer(server, recoveryRuntimeHealthServer{})
-	go func() { _ = server.Serve(listener) }()
-	t.Cleanup(server.Stop)
-
-	endpoint := "unix://" + path
-	client, err := createPostHandoffRuntimeCommandClient(t.Context(), RuntimeControlConfig{Endpoint: endpoint, Token: token}, root)
-	require.NoError(t, err)
-	closer, ok := client.(interface{ Close() error })
-	require.True(t, ok)
-	t.Cleanup(func() { require.NoError(t, closer.Close()) })
-	health, ok := client.(out.RuntimeHealthClient)
-	require.True(t, ok)
-	require.NoError(t, health.PingRuntime(t.Context()), "the generated recovery token must authenticate over the host socket")
-
-	_, err = createRuntimeCommandClient(t.Context(), RuntimeControlConfig{Endpoint: endpoint, Token: token})
-	require.Error(t, err, "the generic runtime client must not accept host migration or engine sockets")
-}
-
-func TestPrivateBootstrapRuntimeClientWaitsForConnectableSocketBeforeConstruction(t *testing.T) {
-	path := newStaleValidatedRuntimeSocketPath(t)
-	var listener net.Listener
-	var server *grpc.Server
-	retryCalls := 0
-	retry := func(context.Context) error {
-		retryCalls++
-		require.NoError(t, os.Remove(path))
-		var err error
-		listener, err = net.Listen("unix", path)
-		require.NoError(t, err)
-		server = newAuthenticatedBootstrapRuntimeServer("bootstrap-token")
-		go func() { _ = server.Serve(listener) }()
-		return nil
-	}
-
-	client, err := createPrivateBootstrapRuntimeCommandClientWithRetry(t.Context(), RuntimeControlConfig{Token: "bootstrap-token"}, "passthrough:///runtime-bootstrap", func(ctx context.Context) (net.Conn, error) {
-		return dialValidatedRuntimeSocket(ctx, path)
-	}, retry)
-	require.NoError(t, err)
-	require.Equal(t, 1, retryCalls, "the missing endpoint must be observed before client construction")
-	closer := client.(io.Closer)
-	t.Cleanup(func() {
-		require.NoError(t, closer.Close())
-		server.Stop()
-		_ = listener.Close()
-	})
-
-	probe, ok := client.(out.RuntimeEnvironmentProbe)
-	require.True(t, ok)
-	result, err := probe.ProbeRuntimeEnvironment(t.Context())
-	require.NoError(t, err)
-	assert.True(t, result.APIReachable)
-	assert.True(t, result.Rootless)
-}
-
-func TestPrivateBootstrapRuntimeClientDoesNotRetryTransportAfterReadinessBarrier(t *testing.T) {
-	var dialCalls atomic.Int32
-	var readinessPeer net.Conn
-	dial := func(context.Context) (net.Conn, error) {
-		if dialCalls.Add(1) == 1 {
-			server, client := net.Pipe()
-			readinessPeer = server
-			return client, nil
-		}
-		return nil, errPrivateRuntimeTransportUnavailable
-	}
-
-	client, err := createPrivateBootstrapRuntimeCommandClientWithRetry(t.Context(), RuntimeControlConfig{Token: "bootstrap-token"}, "passthrough:///runtime-bootstrap", dial, func(context.Context) error {
-		t.Fatal("the connectable endpoint must satisfy the application readiness barrier")
-		return nil
-	})
-	require.NoError(t, err)
-	require.NoError(t, readinessPeer.Close())
-	closer := client.(io.Closer)
-	t.Cleanup(func() { require.NoError(t, closer.Close()) })
-
-	ctx, cancel := context.WithTimeout(t.Context(), time.Second)
-	defer cancel()
-	_, err = client.(out.RuntimeEnvironmentProbe).ProbeRuntimeEnvironment(ctx)
-	require.Error(t, err)
-	assert.Equal(t, codes.Unavailable, status.Code(err))
-	assert.Equal(t, int32(2), dialCalls.Load(), "post-barrier RPCs must dial once without a second transport-readiness retry loop")
-}
-
-func TestPrivateBootstrapRuntimeClientRejectsMissingTokenBeforeTransportWait(t *testing.T) {
-	_, err := createPrivateBootstrapRuntimeCommandClientWithRetry(t.Context(), RuntimeControlConfig{}, "passthrough:///runtime-bootstrap", func(context.Context) (net.Conn, error) {
-		t.Fatal("invalid credentials must fail before transport readiness")
-		return nil, nil
-	}, func(context.Context) error {
-		t.Fatal("invalid credentials must not enter readiness retries")
-		return nil
-	})
-	require.EqualError(t, err, "private runtime authentication token is required")
-}
-
-func TestPrivateBootstrapRuntimeClientResolvesTokenBeforeTransportWait(t *testing.T) {
-	path := newStaleValidatedRuntimeSocketPath(t)
-	const (
-		tokenEnv = "GORDON_TEST_RUNTIME_TOKEN"
-		token    = "resolved-bootstrap-token"
-	)
-	t.Setenv(tokenEnv, token)
-	var listener net.Listener
-	var server *grpc.Server
-	retry := func(context.Context) error {
-		t.Setenv(tokenEnv, "")
-		require.NoError(t, os.Remove(path))
-		var err error
-		listener, err = net.Listen("unix", path)
-		require.NoError(t, err)
-		server = newAuthenticatedBootstrapRuntimeServer(token)
-		go func() { _ = server.Serve(listener) }()
-		return nil
-	}
-
-	client, err := createPrivateBootstrapRuntimeCommandClientWithRetry(t.Context(), RuntimeControlConfig{TokenEnv: tokenEnv}, "passthrough:///runtime-bootstrap", func(ctx context.Context) (net.Conn, error) {
-		return dialValidatedRuntimeSocket(ctx, path)
-	}, retry)
-	require.NoError(t, err, "client construction must retain the token resolved before transport readiness")
-	closer := client.(io.Closer)
-	t.Cleanup(func() {
-		require.NoError(t, closer.Close())
-		server.Stop()
-		_ = listener.Close()
-	})
-
-	result, err := client.(out.RuntimeEnvironmentProbe).ProbeRuntimeEnvironment(t.Context())
-	require.NoError(t, err, "the retained token must authenticate protocol calls")
-	assert.True(t, result.APIReachable)
-}
-
-func TestWaitForPrivateRuntimeTransportClosesReadinessProbe(t *testing.T) {
-	server, client := net.Pipe()
-	closed := make(chan error, 1)
-	go func() {
-		buffer := make([]byte, 1)
-		_, err := server.Read(buffer)
-		closed <- err
-	}()
-	t.Cleanup(func() { require.NoError(t, server.Close()) })
-
-	err := waitForPrivateRuntimeTransport(t.Context(), func(context.Context) (net.Conn, error) {
-		return client, nil
-	}, func(context.Context) error {
-		t.Fatal("a successful readiness probe must not retry")
-		return nil
-	})
-	require.NoError(t, err)
-	require.ErrorIs(t, <-closed, io.EOF)
-}
-
-func TestWaitForPrivateRuntimeTransportRetriesInspectionUntilSocketIsPublished(t *testing.T) {
-	root, err := os.MkdirTemp("/tmp", "gordon-inspection-")
-	require.NoError(t, err)
-	t.Cleanup(func() { require.NoError(t, os.RemoveAll(root)) })
-	dataDir := filepath.Join(root, "data")
-	migrationDir := filepath.Join(dataDir, "migration")
-	candidateDir := filepath.Join(migrationDir, "fixture")
-	for _, dir := range []string{dataDir, migrationDir, candidateDir} {
-		require.NoError(t, os.Mkdir(dir, 0o700))
-	}
-	path := filepath.Join(candidateDir, bootstrapRuntimeSocketName)
-	_, initialErr := dialValidatedRuntimeSocket(t.Context(), path)
-	require.Error(t, initialErr)
-	assert.Equal(t, codes.PermissionDenied, status.Code(initialErr))
-	assert.Contains(t, initialErr.Error(), "category=inspection_failure")
-	assert.NotContains(t, initialErr.Error(), path)
-
-	var listener net.Listener
-	t.Cleanup(func() {
-		if listener != nil {
-			_ = listener.Close()
-		}
-	})
-	dialCalls := 0
-	retryCalls := 0
-	err = waitForPrivateRuntimeTransport(t.Context(), func(ctx context.Context) (net.Conn, error) {
-		dialCalls++
-		return dialValidatedRuntimeSocket(ctx, path)
-	}, func(context.Context) error {
-		retryCalls++
-		if retryCalls < 3 {
-			return nil
-		}
-		var listenErr error
-		listener, listenErr = net.Listen("unix", path)
-		require.NoError(t, listenErr)
-		require.NoError(t, os.Chmod(path, 0o600))
-		info, statErr := os.Lstat(path)
-		require.NoError(t, statErr)
-		assert.NotZero(t, info.Mode()&os.ModeSocket)
-		assert.Equal(t, os.FileMode(0o600), info.Mode().Perm())
-		assert.Equal(t, uint32(os.Geteuid()), info.Sys().(*syscall.Stat_t).Uid)
-		return nil
-	})
-	require.NoError(t, err)
-	require.NoError(t, listener.Close())
-	assert.Equal(t, 4, dialCalls)
-	assert.Equal(t, 3, retryCalls)
-
-	for _, dir := range []string{dataDir, migrationDir, candidateDir} {
-		info, statErr := os.Lstat(dir)
-		require.NoError(t, statErr)
-		assert.Equal(t, os.FileMode(0o700), info.Mode().Perm())
-		assert.Equal(t, uint32(os.Geteuid()), info.Sys().(*syscall.Stat_t).Uid)
-	}
-}
-
-func TestWaitForPrivateRuntimeTransportPersistentInspectionStopsWithCanonicalCategory(t *testing.T) {
-	path := filepath.Join(t.TempDir(), bootstrapRuntimeSocketName)
-	dialCalls := 0
-	retryCalls := 0
-	err := waitForPrivateRuntimeTransport(t.Context(), func(ctx context.Context) (net.Conn, error) {
-		dialCalls++
-		return dialValidatedRuntimeSocket(ctx, path)
-	}, func(context.Context) error {
-		retryCalls++
-		if retryCalls == 3 {
-			return context.DeadlineExceeded
-		}
-		return nil
-	})
-
-	require.ErrorIs(t, err, context.DeadlineExceeded)
-	assert.Equal(t, codes.PermissionDenied, status.Code(err))
-	assert.Contains(t, err.Error(), "category=inspection_failure")
-	assert.NotContains(t, err.Error(), path)
-	assert.Equal(t, 3, dialCalls)
-	assert.Equal(t, 3, retryCalls)
-}
-
-func TestWaitForPrivateRuntimeTransportRerunsSecurityValidationAfterInspection(t *testing.T) {
-	for _, tc := range []struct {
-		name     string
-		category string
-		publish  func(*testing.T, string) string
-	}{
-		{
-			name:     "non-canonical shape",
-			category: "invalid_shape",
-			publish: func(_ *testing.T, root string) string {
-				return filepath.Join(root, "wrong.sock")
-			},
-		},
-		{
-			name:     "symlink ancestor",
-			category: "symlink_ancestor",
-			publish: func(t *testing.T, root string) string {
-				realDir := filepath.Join(root, "real")
-				require.NoError(t, os.Mkdir(realDir, 0o700))
-				linkDir := filepath.Join(root, "linked")
-				require.NoError(t, os.Symlink(realDir, linkDir))
-				return filepath.Join(linkDir, bootstrapRuntimeSocketName)
-			},
-		},
-		{
-			name:     "symlink node",
-			category: "invalid_node",
-			publish: func(t *testing.T, root string) string {
-				candidateDir := filepath.Join(root, "symlink-node")
-				require.NoError(t, os.Mkdir(candidateDir, 0o700))
-				path := filepath.Join(candidateDir, bootstrapRuntimeSocketName)
-				require.NoError(t, os.Symlink(filepath.Join(root, "outside.sock"), path))
-				return path
-			},
-		},
-		{
-			name:     "non-socket node",
-			category: "invalid_node",
-			publish: func(t *testing.T, root string) string {
-				candidateDir := filepath.Join(root, "regular-node")
-				require.NoError(t, os.Mkdir(candidateDir, 0o700))
-				path := filepath.Join(candidateDir, bootstrapRuntimeSocketName)
-				require.NoError(t, os.WriteFile(path, []byte("not a socket"), 0o600))
-				return path
-			},
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			root := t.TempDir()
-			initialDir := filepath.Join(root, "initial")
-			require.NoError(t, os.Mkdir(initialDir, 0o700))
-			path := filepath.Join(initialDir, bootstrapRuntimeSocketName)
-			dialCalls := 0
-			retryCalls := 0
-
-			err := waitForPrivateRuntimeTransport(t.Context(), func(ctx context.Context) (net.Conn, error) {
-				dialCalls++
-				return dialValidatedRuntimeSocket(ctx, path)
-			}, func(context.Context) error {
-				retryCalls++
-				path = tc.publish(t, root)
-				return nil
-			})
-
-			require.Error(t, err)
-			assert.Equal(t, codes.PermissionDenied, status.Code(err))
-			assert.Contains(t, err.Error(), "category="+tc.category)
-			assert.NotContains(t, err.Error(), root)
-			assert.Equal(t, 2, dialCalls)
-			assert.Equal(t, 1, retryCalls)
-		})
-	}
-}
-
-func TestWaitForPrivateRuntimeTransportInspectionCancellationIsPrompt(t *testing.T) {
-	path := filepath.Join(t.TempDir(), bootstrapRuntimeSocketName)
-	ctx, cancel := context.WithCancel(t.Context())
-	waitStarted := make(chan struct{})
-	done := make(chan error, 1)
-	go func() {
-		done <- waitForPrivateRuntimeTransport(ctx, func(ctx context.Context) (net.Conn, error) {
-			return dialValidatedRuntimeSocket(ctx, path)
-		}, func(waitCtx context.Context) error {
-			close(waitStarted)
-			<-waitCtx.Done()
-			return waitCtx.Err()
-		})
-	}()
-
-	<-waitStarted
-	cancel()
-	select {
-	case err := <-done:
-		require.ErrorIs(t, err, context.Canceled)
-		assert.NotContains(t, err.Error(), path)
-	case <-time.After(250 * time.Millisecond):
-		t.Fatal("inspection retry did not stop promptly after cancellation")
-	}
-}
-
-func TestWaitForPrivateRuntimeTransportMissingUntilDeadline(t *testing.T) {
-	ctx, cancel := context.WithDeadline(t.Context(), time.Now().Add(-time.Second))
-	defer cancel()
-	dialCalls := 0
-	err := waitForPrivateRuntimeTransport(ctx, func(context.Context) (net.Conn, error) {
-		dialCalls++
-		return nil, errPrivateRuntimeTransportUnavailable
-	}, func(context.Context) error {
-		t.Fatal("an expired deadline must stop before retry")
-		return nil
-	})
-	require.ErrorIs(t, err, context.DeadlineExceeded)
-	assert.Equal(t, 1, dialCalls)
-}
-
-func TestWaitForPrivateRuntimeTransportStopsPromptlyWhenCanceled(t *testing.T) {
-	ctx, cancel := context.WithCancel(t.Context())
-	waitStarted := make(chan struct{})
-	done := make(chan error, 1)
-	go func() {
-		done <- waitForPrivateRuntimeTransport(ctx, func(context.Context) (net.Conn, error) {
-			return nil, errPrivateRuntimeTransportUnavailable
-		}, func(waitCtx context.Context) error {
-			close(waitStarted)
-			<-waitCtx.Done()
-			return waitCtx.Err()
-		})
-	}()
-	<-waitStarted
-	cancel()
-	require.ErrorIs(t, <-done, context.Canceled)
-}
-
-func TestDialValidatedRuntimeSocketReportsValueFreeValidationCategories(t *testing.T) {
-	tests := []struct {
-		name         string
-		category     string
-		prepare      func(*testing.T) string
-		connectError error
-		code         codes.Code
-	}{
-		{
-			name:     "relative shape",
-			category: "invalid_shape",
-			prepare: func(*testing.T) string {
-				return filepath.Join("private-fixture", bootstrapRuntimeSocketName)
-			},
-			code: codes.PermissionDenied,
-		},
-		{
-			name:     "non-canonical shape",
-			category: "invalid_shape",
-			prepare: func(t *testing.T) string {
-				return t.TempDir() + string(filepath.Separator) + "nested" + string(filepath.Separator) + ".." + string(filepath.Separator) + bootstrapRuntimeSocketName
-			},
-			code: codes.PermissionDenied,
-		},
-		{
-			name:     "invalid basename",
-			category: "invalid_shape",
-			prepare: func(t *testing.T) string {
-				return filepath.Join(t.TempDir(), "wrong.sock")
-			},
-			code: codes.PermissionDenied,
-		},
-		{
-			name:     "symlink ancestor",
-			category: "symlink_ancestor",
-			prepare: func(t *testing.T) string {
-				root := t.TempDir()
-				ancestor := filepath.Join(root, "private-ancestor")
-				realDirectory := t.TempDir()
-				require.NoError(t, os.Symlink(realDirectory, ancestor))
-				return filepath.Join(ancestor, bootstrapRuntimeSocketName)
-			},
-			code: codes.PermissionDenied,
-		},
-		{
-			name:     "missing node inspection",
-			category: "inspection_failure",
-			prepare: func(t *testing.T) string {
-				return filepath.Join(t.TempDir(), bootstrapRuntimeSocketName)
-			},
-			code: codes.PermissionDenied,
-		},
-		{
-			name:     "symlink node",
-			category: "invalid_node",
-			prepare: func(t *testing.T) string {
-				path := filepath.Join(t.TempDir(), bootstrapRuntimeSocketName)
-				require.NoError(t, os.Symlink(filepath.Join(t.TempDir(), "outside.sock"), path))
-				return path
-			},
-			code: codes.PermissionDenied,
-		},
-		{
-			name:     "non-socket node",
-			category: "invalid_node",
-			prepare: func(t *testing.T) string {
-				path := filepath.Join(t.TempDir(), bootstrapRuntimeSocketName)
-				require.NoError(t, os.WriteFile(path, []byte("not a socket"), 0o600))
-				return path
-			},
-			code: codes.PermissionDenied,
-		},
-		{
-			name:         "connect permission",
-			category:     "connect_permission",
-			prepare:      newValidatedRuntimeSocketPath,
-			connectError: os.ErrPermission,
-			code:         codes.PermissionDenied,
-		},
-		{
-			name:         "connect unavailable",
-			category:     "connect_unavailable",
-			prepare:      newValidatedRuntimeSocketPath,
-			connectError: syscall.ECONNREFUSED,
-			code:         codes.Unknown,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			path := tc.prepare(t)
-			dialCalls := 0
-			_, err := dialValidatedRuntimeSocketWithDialer(t.Context(), path, func(context.Context, string, string) (net.Conn, error) {
-				dialCalls++
-				if tc.category == "connect_permission" {
-					return nil, fmt.Errorf("endpoint=%s uid=12345 mode=0777 value=private-value: %w", path, tc.connectError)
-				}
-				return nil, tc.connectError
-			})
-			require.Error(t, err)
-			assert.Equal(t, tc.code, status.Code(err))
-			assert.Contains(t, err.Error(), "category="+tc.category)
-			assert.NotContains(t, err.Error(), path)
-			assert.NotContains(t, err.Error(), "uid=12345")
-			assert.NotContains(t, err.Error(), "mode=0777")
-			assert.NotContains(t, err.Error(), "value=private-value")
-			if tc.connectError == nil {
-				assert.Zero(t, dialCalls)
-			} else {
-				assert.Equal(t, 1, dialCalls)
-			}
-		})
-	}
-}
-
-func TestValidatePrivateRuntimeSocketPathReportsValueFreeInspectionFailures(t *testing.T) {
-	rootInfo, err := os.Lstat(string(filepath.Separator))
-	require.NoError(t, err)
-
-	for _, tc := range []struct {
-		name   string
-		path   string
-		failAt string
-	}{
-		{
-			name:   "ancestor permission failure",
-			path:   filepath.Join(string(filepath.Separator), "private-ancestor", bootstrapRuntimeSocketName),
-			failAt: filepath.Join(string(filepath.Separator), "private-ancestor"),
-		},
-		{
-			name:   "node permission failure",
-			path:   filepath.Join(string(filepath.Separator), bootstrapRuntimeSocketName),
-			failAt: filepath.Join(string(filepath.Separator), bootstrapRuntimeSocketName),
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			const maliciousText = "inspection token=do-not-leak"
-			err := validatePrivateRuntimeSocketPath(tc.path, func(path string) (os.FileInfo, error) {
-				if path == tc.failAt {
-					return nil, fmt.Errorf("inspect path=%s %s: %w", path, maliciousText, os.ErrPermission)
-				}
-				return rootInfo, nil
-			})
-			require.Error(t, err)
-			assert.Equal(t, codes.PermissionDenied, status.Code(err))
-			assert.Contains(t, err.Error(), "category=inspection_failure")
-			assert.NotContains(t, err.Error(), tc.path)
-			assert.NotContains(t, err.Error(), maliciousText)
-		})
-	}
-}
-
-func TestInspectRuntimeSocketAncestorsDefaultFailClosed(t *testing.T) {
-	err := privateRuntimeTransportErrorForSocketAncestors(runtimeSocketAncestorResult(99))
-	require.Error(t, err)
-	assert.Contains(t, err.Error(), "category=inspection_failure")
-}
-
-func TestPrivateBootstrapRuntimeClientReportsBoundedConnectUnavailable(t *testing.T) {
-	path := newValidatedRuntimeSocketPath(t)
-	_, err := createPrivateBootstrapRuntimeCommandClientWithRetry(t.Context(), RuntimeControlConfig{Token: "bootstrap-token"}, "passthrough:///runtime-bootstrap", func(ctx context.Context) (net.Conn, error) {
-		return dialValidatedRuntimeSocketWithDialer(ctx, path, func(context.Context, string, string) (net.Conn, error) {
-			return nil, wrappedUnixConnectError(syscall.ECONNREFUSED)
-		})
-	}, func(context.Context) error {
-		return context.DeadlineExceeded
-	})
-	require.ErrorIs(t, err, context.DeadlineExceeded)
-	assert.Contains(t, err.Error(), "wait for private runtime transport")
-	assert.Contains(t, err.Error(), "category=connect_unavailable")
-	assert.NotContains(t, err.Error(), path)
-}
-
-func TestWaitForPrivateRuntimeTransportRejectsInvalidTransportsWithoutRetry(t *testing.T) {
-	root := t.TempDir()
-	for _, tc := range []struct {
-		name    string
-		path    string
-		prepare func(string) error
-	}{
-		{name: "invalid path", path: filepath.Join(root, "wrong.sock")},
-		{name: "regular file", path: filepath.Join(root, "regular", bootstrapRuntimeSocketName), prepare: func(path string) error {
-			if err := os.MkdirAll(filepath.Dir(path), 0o700); err != nil {
-				return err
-			}
-			return os.WriteFile(path, []byte("not a socket"), 0o600)
-		}},
-		{name: "symlink", path: filepath.Join(root, "symlink", bootstrapRuntimeSocketName), prepare: func(path string) error {
-			if err := os.MkdirAll(filepath.Dir(path), 0o700); err != nil {
-				return err
-			}
-			return os.Symlink(filepath.Join(root, "outside.sock"), path)
-		}},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			if tc.prepare != nil {
-				require.NoError(t, tc.prepare(tc.path))
-			}
-			err := waitForPrivateRuntimeTransport(t.Context(), func(ctx context.Context) (net.Conn, error) {
-				return dialValidatedRuntimeSocket(ctx, tc.path)
-			}, func(context.Context) error {
-				t.Fatal("invalid transport must fail closed without retry")
-				return nil
-			})
-			require.Error(t, err)
-			assert.Equal(t, codes.PermissionDenied, status.Code(err))
-		})
-	}
-}
-
-func TestPrivateTransportErrorClassifierUnifiesCategories(t *testing.T) {
-	const maliciousText = "path=/private/runtime.sock uid=12345 token=do-not-leak"
-
-	terminalCategories := []privateRuntimeTransportErrorCategory{
-		privateRuntimeTransportInvalidShape,
-		privateRuntimeTransportSymlinkAncestor,
-		privateRuntimeTransportInvalidNode,
-		privateRuntimeTransportConnectPermission,
-		privateRuntimeTransportUnvalidatedFailure,
-	}
-	for _, category := range terminalCategories {
-		t.Run("terminal_"+string(category), func(t *testing.T) {
-			err := privateRuntimeTransportValidationError(category)
-			require.False(t, isTransientPrivateRuntimeTransportError(err))
-			gotCategory, ok := privateRuntimeTransportValidationCategory(err)
-			require.True(t, ok)
-			assert.Equal(t, category, gotCategory)
-			assert.Equal(t, codes.PermissionDenied, status.Code(err))
-			assert.Contains(t, err.Error(), "category="+string(category))
-			assert.NotContains(t, err.Error(), maliciousText)
-
-			wrapped := fmt.Errorf("%s: %w", maliciousText, err)
-			gotCategory, ok = privateRuntimeTransportValidationCategory(wrapped)
-			require.True(t, ok)
-			assert.Equal(t, category, gotCategory)
-			assert.Equal(t, codes.PermissionDenied, status.Code(wrapped))
-			var transportErr *privateRuntimeTransportError
-			require.True(t, errors.As(wrapped, &transportErr))
-			assert.NotContains(t, transportErr.Error(), maliciousText)
-		})
-	}
-
-	t.Run("transient_connect_unavailable", func(t *testing.T) {
-		err := errPrivateRuntimeTransportUnavailable
-		require.True(t, isTransientPrivateRuntimeTransportError(err))
-		gotCategory, ok := privateRuntimeTransportValidationCategory(err)
-		require.True(t, ok)
-		assert.Equal(t, privateRuntimeTransportConnectUnavailable, gotCategory)
-		assert.Equal(t, codes.Unknown, status.Code(err))
-		assert.Contains(t, err.Error(), "category=connect_unavailable")
-	})
-
-	t.Run("transient_inspection_failure", func(t *testing.T) {
-		err := privateRuntimeTransportValidationError(privateRuntimeTransportInspectionFailure)
-		require.True(t, isTransientPrivateRuntimeTransportError(err))
-		gotCategory, ok := privateRuntimeTransportValidationCategory(err)
-		require.True(t, ok)
-		assert.Equal(t, privateRuntimeTransportInspectionFailure, gotCategory)
-		assert.Equal(t, codes.PermissionDenied, status.Code(err))
-	})
-
-	t.Run("bounded_unavailable_deadline", func(t *testing.T) {
-		err := privateRuntimeTransportBoundedDeadlineError(privateRuntimeTransportConnectUnavailable)
-		require.ErrorIs(t, err, context.DeadlineExceeded)
-		require.True(t, errors.Is(err, errPrivateRuntimeTransportUnavailable))
-		gotCategory, ok := privateRuntimeTransportValidationCategory(err)
-		require.True(t, ok)
-		assert.Equal(t, privateRuntimeTransportConnectUnavailable, gotCategory)
-		assert.Contains(t, err.Error(), "category=connect_unavailable")
-		assert.NotContains(t, err.Error(), maliciousText)
-
-		wrapped := fmt.Errorf("%s: %w", maliciousText, err)
-		require.ErrorIs(t, wrapped, context.DeadlineExceeded)
-		require.True(t, errors.Is(wrapped, errPrivateRuntimeTransportUnavailable))
-		gotCategory, ok = privateRuntimeTransportValidationCategory(wrapped)
-		require.True(t, ok)
-		assert.Equal(t, privateRuntimeTransportConnectUnavailable, gotCategory)
-		var transportErr *privateRuntimeTransportError
-		require.True(t, errors.As(wrapped, &transportErr))
-		assert.NotContains(t, transportErr.Error(), maliciousText)
-	})
-
-	t.Run("bounded_inspection_deadline", func(t *testing.T) {
-		err := privateRuntimeTransportBoundedDeadlineError(privateRuntimeTransportInspectionFailure)
-		require.ErrorIs(t, err, context.DeadlineExceeded)
-		gotCategory, ok := privateRuntimeTransportValidationCategory(err)
-		require.True(t, ok)
-		assert.Equal(t, privateRuntimeTransportInspectionFailure, gotCategory)
-		assert.Equal(t, codes.PermissionDenied, status.Code(err))
-		assert.Contains(t, err.Error(), "category=inspection_failure")
-		assert.NotContains(t, err.Error(), maliciousText)
-	})
-
-	t.Run("deadline_classifier_from_dial_error", func(t *testing.T) {
-		inspectionErr := privateRuntimeTransportValidationError(privateRuntimeTransportInspectionFailure)
-		unavailableErr := errPrivateRuntimeTransportUnavailable
-
-		inspectionBounded := boundedPrivateRuntimeTransportDeadlineError(inspectionErr)
-		require.ErrorIs(t, inspectionBounded, context.DeadlineExceeded)
-		assert.Equal(t, codes.PermissionDenied, status.Code(inspectionBounded))
-
-		unavailableBounded := boundedPrivateRuntimeTransportDeadlineError(unavailableErr)
-		require.ErrorIs(t, unavailableBounded, context.DeadlineExceeded)
-		require.True(t, errors.Is(unavailableBounded, errPrivateRuntimeTransportUnavailable))
-	})
-
-	t.Run("retry_context_normalization", func(t *testing.T) {
-		err := waitForPrivateRuntimeTransport(t.Context(), func(context.Context) (net.Conn, error) {
-			return nil, fmt.Errorf("dial %s: %w", maliciousText, errPrivateRuntimeTransportUnavailable)
-		}, func(context.Context) error {
-			return context.DeadlineExceeded
-		})
-		require.ErrorIs(t, err, context.DeadlineExceeded)
-		require.True(t, errors.Is(err, errPrivateRuntimeTransportUnavailable))
-		assert.NotContains(t, err.Error(), maliciousText)
-	})
-}
-
-func TestWaitForPrivateRuntimeTransportNormalizesWrappedValidationStatus(t *testing.T) {
-	const maliciousText = "caller path=/private/runtime.sock token=do-not-leak"
-	for _, tc := range []struct {
-		name     string
-		category privateRuntimeTransportErrorCategory
-		want     privateRuntimeTransportErrorCategory
-	}{
-		{name: "known category", category: privateRuntimeTransportInvalidNode, want: privateRuntimeTransportInvalidNode},
-		{name: "unknown category", category: privateRuntimeTransportErrorCategory(maliciousText), want: privateRuntimeTransportUnvalidatedFailure},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			wrapped := fmt.Errorf("%s: %w", maliciousText, &privateRuntimeTransportError{category: tc.category})
-			err := waitForPrivateRuntimeTransport(t.Context(), func(context.Context) (net.Conn, error) {
-				return nil, wrapped
-			}, func(context.Context) error {
-				t.Fatal("terminal validation status must not retry")
-				return nil
-			})
-
-			require.Error(t, err)
-			assert.Equal(t, codes.PermissionDenied, status.Code(err))
-			assert.Contains(t, err.Error(), "category="+string(tc.want))
-			assert.NotContains(t, err.Error(), maliciousText)
-		})
-	}
-}
-
-func TestWaitForPrivateRuntimeTransportReportsValueFreeTerminalCategories(t *testing.T) {
-	for _, tc := range []struct {
-		name     string
-		dialErr  error
-		category string
-	}{
-		{
-			name:     "permission failure",
-			dialErr:  fmt.Errorf("endpoint=private uid=12345 mode=0777 value=secret: %w", os.ErrPermission),
-			category: "connect_permission",
-		},
-		{
-			name:     "unvalidated failure",
-			dialErr:  fmt.Errorf("endpoint=private uid=12345 mode=0777 value=secret: %w", syscall.EIO),
-			category: "unvalidated_failure",
-		},
-		{
-			name:     "unclassified permission status",
-			dialErr:  status.Error(codes.PermissionDenied, "endpoint=private uid=12345 mode=0777 value=secret"),
-			category: "unvalidated_failure",
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			err := waitForPrivateRuntimeTransport(t.Context(), func(context.Context) (net.Conn, error) {
-				return nil, tc.dialErr
-			}, func(context.Context) error {
-				t.Fatal("terminal failure must fail closed without retry")
-				return nil
-			})
-			require.Error(t, err)
-			assert.Equal(t, codes.PermissionDenied, status.Code(err))
-			assert.Contains(t, err.Error(), "category="+tc.category)
-			assert.NotContains(t, err.Error(), "endpoint=private")
-			assert.NotContains(t, err.Error(), "uid=12345")
-			assert.NotContains(t, err.Error(), "mode=0777")
-			assert.NotContains(t, err.Error(), "value=secret")
-		})
-	}
-}
-
-func TestWaitForPrivateRuntimeTransportDoesNotRetryUnvalidatedConnectErrors(t *testing.T) {
-	for _, connectErr := range []error{syscall.ENOENT, syscall.ECONNREFUSED, syscall.EAGAIN} {
-		t.Run(connectErr.Error(), func(t *testing.T) {
-			err := waitForPrivateRuntimeTransport(t.Context(), func(context.Context) (net.Conn, error) {
-				return nil, wrappedUnixConnectError(connectErr)
-			}, func(context.Context) error {
-				t.Fatal("a connect error must not retry without validated-socket classification")
-				return nil
-			})
-			require.Error(t, err)
-			assert.Equal(t, codes.PermissionDenied, status.Code(err))
-		})
-	}
-}
-
-func TestWaitForPrivateRuntimeTransportRetriesValidatedSocketConnectFailures(t *testing.T) {
-	path := newValidatedRuntimeSocketPath(t)
-	for _, tc := range []struct {
-		name string
-		err  error
-	}{
-		{name: "again", err: syscall.EAGAIN},
-		{name: "interrupted", err: syscall.EINTR},
-		{name: "invalid argument", err: syscall.EINVAL},
-		{name: "no such device or address", err: syscall.ENXIO},
-		{name: "connection reset", err: syscall.ECONNRESET},
-		{name: "connection refused", err: syscall.ECONNREFUSED},
-		{name: "socket disappeared", err: syscall.ENOENT},
-		{name: "unknown non-permission failure", err: syscall.EIO},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			dialCalls := 0
-			retryCalls := 0
-			var peer net.Conn
-			err := waitForPrivateRuntimeTransport(t.Context(), func(ctx context.Context) (net.Conn, error) {
-				return dialValidatedRuntimeSocketWithDialer(ctx, path, func(context.Context, string, string) (net.Conn, error) {
-					dialCalls++
-					if dialCalls == 1 {
-						return nil, wrappedUnixConnectError(tc.err)
-					}
-					server, client := net.Pipe()
-					peer = server
-					return client, nil
-				})
-			}, func(context.Context) error {
-				retryCalls++
-				return nil
-			})
-			require.NoError(t, err)
-			require.NoError(t, peer.Close())
-			assert.Equal(t, 2, dialCalls)
-			assert.Equal(t, 1, retryCalls)
-		})
-	}
-}
-
-func TestWaitForPrivateRuntimeTransportRejectsValidatedSocketSecurityErrorsImmediately(t *testing.T) {
-	path := newValidatedRuntimeSocketPath(t)
-	for _, permissionErr := range []error{syscall.EACCES, syscall.EPERM} {
-		t.Run(permissionErr.Error(), func(t *testing.T) {
-			dialCalls := 0
-			err := waitForPrivateRuntimeTransport(t.Context(), func(ctx context.Context) (net.Conn, error) {
-				return dialValidatedRuntimeSocketWithDialer(ctx, path, func(context.Context, string, string) (net.Conn, error) {
-					dialCalls++
-					return nil, wrappedUnixConnectError(permissionErr)
-				})
-			}, func(context.Context) error {
-				t.Fatal("permission failures from a validated socket must not retry")
-				return nil
-			})
-			require.Error(t, err)
-			assert.Equal(t, codes.PermissionDenied, status.Code(err))
-			assert.Equal(t, 1, dialCalls)
-		})
-	}
-}
-
-func TestWaitForPrivateRuntimeTransportRejectsWrappedConnectContextErrorsImmediately(t *testing.T) {
-	path := newValidatedRuntimeSocketPath(t)
-	for _, contextErr := range []error{context.Canceled, context.DeadlineExceeded} {
-		t.Run(contextErr.Error(), func(t *testing.T) {
-			err := waitForPrivateRuntimeTransport(t.Context(), func(ctx context.Context) (net.Conn, error) {
-				return dialValidatedRuntimeSocketWithDialer(ctx, path, func(context.Context, string, string) (net.Conn, error) {
-					return nil, wrappedUnixConnectError(contextErr)
-				})
-			}, func(context.Context) error {
-				t.Fatal("connect context failures must not retry")
-				return nil
-			})
-			require.ErrorIs(t, err, contextErr)
-		})
-	}
-}
-
-func TestWaitForPrivateRuntimeTransportNormalizesRetryContextCauses(t *testing.T) {
-	for _, tc := range []struct {
-		name              string
-		cause             error
-		wantUnavailableIs bool
-	}{
-		{name: "deadline", cause: context.DeadlineExceeded, wantUnavailableIs: true},
-		{name: "cancellation", cause: context.Canceled},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			const maliciousText = "retry path=/private/runtime.sock token=do-not-leak"
-			err := waitForPrivateRuntimeTransport(t.Context(), func(context.Context) (net.Conn, error) {
-				return nil, fmt.Errorf("dial path=/another/private.sock: %w", errPrivateRuntimeTransportUnavailable)
-			}, func(context.Context) error {
-				return fmt.Errorf("%s: %w", maliciousText, tc.cause)
-			})
-
-			require.ErrorIs(t, err, tc.cause)
-			assert.Equal(t, tc.wantUnavailableIs, errors.Is(err, errPrivateRuntimeTransportUnavailable))
-			assert.NotContains(t, err.Error(), maliciousText)
-			assert.NotContains(t, err.Error(), "/another/private.sock")
-		})
-	}
-}
-
-func TestWaitForPrivateRuntimeTransportValidatedConnectFailureStopsAtDeadline(t *testing.T) {
-	path := newValidatedRuntimeSocketPath(t)
-	dialCalls := 0
-	retryCalls := 0
-	err := waitForPrivateRuntimeTransport(t.Context(), func(ctx context.Context) (net.Conn, error) {
-		dialCalls++
-		return dialValidatedRuntimeSocketWithDialer(ctx, path, func(context.Context, string, string) (net.Conn, error) {
-			return nil, wrappedUnixConnectError(syscall.EAGAIN)
-		})
-	}, func(context.Context) error {
-		retryCalls++
-		if retryCalls == 3 {
-			return context.DeadlineExceeded
-		}
-		return nil
-	})
-	require.ErrorIs(t, err, context.DeadlineExceeded)
-	assert.Equal(t, 3, dialCalls)
-	assert.Equal(t, 3, retryCalls)
-}
-
-func newValidatedRuntimeSocketPath(t *testing.T) string {
-	t.Helper()
-	path := filepath.Join(t.TempDir(), bootstrapRuntimeSocketName)
-	listener, err := net.Listen("unix", path)
-	require.NoError(t, err)
-	t.Cleanup(func() { require.NoError(t, listener.Close()) })
-	return path
-}
-
-func newStaleValidatedRuntimeSocketPath(t *testing.T) string {
-	t.Helper()
-	path := filepath.Join(t.TempDir(), bootstrapRuntimeSocketName)
-	listener, err := net.ListenUnix("unix", &net.UnixAddr{Name: path, Net: "unix"})
-	require.NoError(t, err)
-	listener.SetUnlinkOnClose(false)
-	require.NoError(t, listener.Close())
-	info, err := os.Lstat(path)
-	require.NoError(t, err)
-	require.NotZero(t, info.Mode()&os.ModeSocket)
-	t.Cleanup(func() {
-		if err := os.Remove(path); err != nil && !os.IsNotExist(err) {
-			t.Errorf("remove stale Unix socket: %v", err)
-		}
-	})
-	return path
-}
-
-func wrappedUnixConnectError(err error) error {
-	return fmt.Errorf("outer connect wrapper: %w", &net.OpError{
-		Op:  "dial",
-		Net: "unix",
-		Err: fmt.Errorf("syscall connect: %w", err),
-	})
 }
 
 func TestPrivateNonBootstrapRuntimeClientRemainsFailFast(t *testing.T) {
@@ -1078,33 +109,6 @@ func TestPostHandoffRuntimeClientRejectsMissingRegularAndSymlinkSockets(t *testi
 			}
 		})
 	}
-}
-
-func newAuthenticatedBootstrapRuntimeServer(token string) *grpc.Server {
-	server := grpc.NewServer(grpc.UnaryInterceptor(func(ctx context.Context, req any, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (any, error) {
-		if authorization := metadata.ValueFromIncomingContext(ctx, "authorization"); len(authorization) != 1 || authorization[0] != "Bearer "+token {
-			return nil, status.Error(codes.Unauthenticated, "runtime authentication failed")
-		}
-		return handler(ctx, req)
-	}))
-	runtimev1.RegisterRuntimeServiceServer(server, bootstrapRuntimeProbeServer{})
-	return server
-}
-
-type bootstrapRuntimeProbeServer struct {
-	runtimev1.UnimplementedRuntimeServiceServer
-}
-
-func (bootstrapRuntimeProbeServer) ProbeEnvironment(context.Context, *runtimev1.ProbeEnvironmentRequest) (*runtimev1.ProbeEnvironmentResponse, error) {
-	return &runtimev1.ProbeEnvironmentResponse{Engine: "podman", Rootless: true, ApiReachable: true}, nil
-}
-
-type recoveryRuntimeHealthServer struct {
-	runtimev1.UnimplementedRuntimeServiceServer
-}
-
-func (recoveryRuntimeHealthServer) GetHealth(context.Context, *runtimev1.GetHealthRequest) (*runtimev1.GetHealthResponse, error) {
-	return &runtimev1.GetHealthResponse{Message: "healthy"}, nil
 }
 
 func TestRuntimeComponentLauncherOwnsOnlySuccessfulHandoffClients(t *testing.T) {
